Fix Mongoose model overwrite by consolidating User model
