--- src/models/User.js
@@ -1,54 +1,0 @@
-import mongoose from 'mongoose';
-import validator from 'validator';
-import bcrypt from 'bcryptjs';
-
-const userSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, 'Please tell us your name']
-  },
-  email: {
-    type: String,
-    required: [true, 'Please provide your email'],
-    unique: true,
-    lowercase: true,
-    validate: [validator.isEmail, 'Please provide a valid email']
-  },
-  phone: {
-    type: String,
-    required: [true, 'Please provide your phone number'],
-    unique: true
-  },
-  password: {
-    type: String,
-    required: [true, 'Please provide a password'],
-    minlength: 8,
-    select: false
-  },
-  address: {
-    type: String,
-    required: [true, 'Please provide your address']
-  },
-  avatar: String,
-  role: {
-    type: String,
-    enum: ['user', 'driver', 'admin'],
-    default: 'user'
-  },
-  active: {
-    type: Boolean,
-    default: true,
-    select: false
-  }
-}, {
-  timestamps: true
-});
-
-// Password hashing middleware
-userSchema.pre('save', async function(next) {
-  if (!this.isModified('password')) return next();
-  this.password = await bcrypt.hash(this.password, 12);
-  next();
-});
-
-export default mongoose.model('User', userSchema);
