+++ src/routes/authRoutes.js
@@ -2,7 +2,7 @@ import express from 'express';
 const router = express.Router();
 import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
-import User from '../models/User.js';
+import User from '../models/userModel.js';
 
 router.post('/register', async (req, res) => {
   // userType: "customer", "driver", "vendor"
